Stay compatible and use AkType.LONG, not INT for smallish integers.

+++ src/main/java/com/akiban/sql/optimizer/plan/TypesTranslation.java
@@ -36,7 +36,7 @@ public final class TypesTranslation {
                 return AkType.YEAR;
             if (typeId.isUnsigned())
                 return AkType.U_INT;
-            return AkType.INT;
+            return AkType.LONG; // Not INT.
         case TypeId.FormatIds.LONGINT_TYPE_ID:
             if (typeId.isUnsigned())
                 return AkType.U_BIGINT;
